Migrates Common library: 1.0.0 > 2.0.0

+++ src/main/java/domain/Invoice.java
@@ -57,7 +57,7 @@ public class Invoice implements IInvoice {
     }
 
     @Override
-    public double getPrice() {
+    public BigDecimal getPrice() {
         throw new NotImplementedException();
     }
